Update sample code
