--- Samples/Windows/ReadMe.md
@@ -1,28 +1,0 @@
-# Prerequisites
-
-- [Visual Studio](https://visualstudio.microsoft.com/zh-hant/downloads/) 2017 or later
-- [OpenCV](https://opencv.org/releases/#:~:text=*-,OpenCV%20%E2%80%93%203.4.16,-2021-10-11) (version 3.4.16 or later)
-
-You can refer to [Setting Guide](https://developer.technexion.com/docs/vizionsdk-sample-build-buide-on-windows) for setting up OpenCV on Visual Studio. 
-# Steps to run the VizionSDK sample code:
-
-## Download the sample code:
-
-Visit the VizionSDK Sample download page: https://github.com/TechNexion-Vision/vizionsdk/Samples and download the latest version of the Sample.
-## Extract the downloaded file:
-
-Extract the contents of the downloaded file, likely named ```VizionSDK_win_24.2.0.1.7z.```
-## Run the sample code:
-
-Navigate to the /x64/Release directory within the extracted files.
-Execute the ```VizionSample.exe``` file.
-## Verify the results:
-
-The sample code will initiate camera streaming.
-Follow the instructions displayed in the command line to interact with the sample code and verify its functionality.
-## Additional information:
-
-For more comprehensive information on the VizionSDK Sample, please refer to the official VizionSDK documentation: VisionSDK documentation: https://developer.technexion.com/docs/vizionsdk-overview.
-Note:
-
-These instructions are specifically for running the VizionSDK sample code on a Windows system. The procedure for running the sample code on a Linux system differs slightly. Please consult the VizionSDK documentation: https://developer.technexion.com/docs/vizionsdk-overview for guidance on running the sample code on Linux.
